move implementation into separate package
