+++ http/src/main/java/net/daporkchop/lib/http/server/ServerTest.java
@@ -26,9 +26,9 @@ import io.netty.channel.socket.nio.NioServerSocketChannel;
 import net.daporkchop.lib.common.util.PorkUtil;
 import net.daporkchop.lib.http.Request;
 import net.daporkchop.lib.http.Response;
-import net.daporkchop.lib.http.codec.ExceptionHandlerServerHTTP;
-import net.daporkchop.lib.http.codec.v1.RequestDecoderHTTP1;
-import net.daporkchop.lib.http.codec.v1.ResponseEncoderHTTP1;
+import net.daporkchop.lib.http.impl.netty.codec.ExceptionHandlerServerHTTP;
+import net.daporkchop.lib.http.impl.netty.codec.v1.RequestDecoderHTTP1;
+import net.daporkchop.lib.http.impl.netty.codec.v1.ResponseEncoderHTTP1;
 import net.daporkchop.lib.http.util.ConnectionState;
 import net.daporkchop.lib.http.util.StatusCodes;
 import net.daporkchop.lib.http.util.exception.GenericHTTPException;
